+++ extensions/zoxide-git-projects/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Zoxide Git Projects Changelog
 
-## [Add alternate application and copy shortcuts] - {PR_MERGE_DATE}
+## [Add alternative application and copy shortcuts] - {PR_MERGE_DATE}
 
-- Added support for selecting an alternate application in preferences for opening with `⌘O`.
+- Added support for selecting an alternative application in preferences for opening with `⌘O`.
 - Added shortcuts to copy path and repository name using `⌘⇧,` and `⌘⇧.` respectively.
 
 ## [Improve Fuzzy Matching] - 2024-11-26